norminetted functions
