+++ maps/bad/color_invalid_number.cub
@@ -3,7 +3,7 @@ SO textures/wolfenstein/purple_stone.xpm
 WE textures/wolfenstein/red_brick.xpm
 EA textures/wolfenstein/wood.xpm
 
-F 2 5 5,2 5 5,2 5 5
+F 2,2,,
 C 225,30,0
 
         1111111111111111111111111
